@@ -1,6 +1,6 @@
 {
   "name": "bloomreach-content-saas-ui-extensions",
-  "version": "1.4.0",
+  "version": "2.0.0",
   "title": "Bloomreach Content SaaS UI Extensions",
   "description": "High performance React template built with lots of powerful MUI (Material-UI) components across multiple product niches for fast & perfect apps development processes",
   "author": {
@@ -9,7 +9,7 @@
   },
   "private": false,
   "dependencies": {
-    "@emotion/cache": "11.7.1",
+    "@emotion/cache": "^11.11.0",
     "@emotion/react": "11.9.0",
     "@emotion/server": "11.4.0",
     "@emotion/styled": "11.8.1",
@@ -17,6 +17,8 @@
     "@mui/icons-material": "5.8.2",
     "@mui/lab": "5.0.0-alpha.84",
     "@mui/material": "5.8.2",
+    "@mui/material-nextjs": "^5.15.11",
+    "@mui/styled-engine": "^5.15.14",
     "@mui/styles": "5.8.0",
     "@mui/x-data-grid": "^5.17.20",
     "@reduxjs/toolkit": "^1.9.2",
@@ -28,7 +30,7 @@
     "date-fns": "2.28.0",
     "dependency-graph": "^0.11.0",
     "gray-matter": "^4.0.3",
-    "next": "^13.4.4",
+    "next": "^14.1.4",
     "next-images": "1.8.4",
     "nprogress": "0.2.0",
     "numeral": "2.0.6",
@@ -65,8 +67,11 @@
     "eslint-plugin-prettier": "4.0.0",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
-    "next-transpile-modules": "9.0.0",
+    "next-transpile-modules": "^10.0.1",
     "prettier": "2.6.2",
     "yalc": "1.0.0-pre.53"
+  },
+  "resolutions": {
+    "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest"
   }
 }
