@@ -3,7 +3,7 @@ const { version } = require('./package.json');
 
 module.exports = {
   publicRuntimeConfig: {
-    version,
+    version
   },
   async redirects() {
     return [
